Master

+++ cron_dpo.php
@@ -137,16 +137,90 @@ if($res->access_token){
 $data=$datas->apiDVOResponse->eaiBody->data;
 
 //var_dump($data);
-$rs=exec_qry($conn,"truncate table tm_dpo");
+
+if(date("H")=="8") $rs=exec_qry($conn,"truncate table tm_dpo");
+
 for($i=0;$i<count($data);$i++){
-	$sql="insert ignore into tm_dpo values ('".($i+1)."','".$data[$i]->LOCATION_ID."','".$data[$i]->LOCATION."','".$data[$i]->STATUS."','$now','N','sys')";
+	if($data[$i]->STATUS=='Up'){
+		//$sql="delete from tm_dpo where oid='".$data[$i]->LOCATION_ID."'";
+		$sql="update tm_dpo set stts='".$data[$i]->STATUS."',lastupd=NOW() WHERE oid='".$data[$i]->LOCATION_ID."'";
+	}else{
+		$sql="insert ignore into tm_dpo (oid,oname,stts) values ('".$data[$i]->LOCATION_ID."','".$data[$i]->LOCATION."','".$data[$i]->STATUS."')";
+		$rs=exec_qry($conn,$sql);
+		$sql="update tm_dpo set ctr=ctr+1,lastupd=NOW() WHERE oid='".$data[$i]->LOCATION_ID."'";
+	}
 	//echo $sql."<br />";
 	$rs=exec_qry($conn,$sql);
 }
+
+//$rs=exec_qry($conn,"update tm_dpo set ctr=ctr+1,lastupd=NOW()");
+
+//delete aja yg gaakan diproses
+$sql="delete from tm_dpo where oid IN (SELECT i FROM tm_tickets WHERE s<>'closed' AND (st='wifi station' OR typ IN ('psb','relokasi'))) OR 
+	oid IN (SELECT oid FROM tm_outlets WHERE TIME(wibstart)>TIME(NOW()))";
+$rs=exec_qry($conn,$sql);
 	
+	///create ticket
+	$sql="select *,time(lastupd) as tm, DATE_FORMAT(DATE_SUB(crtd, INTERVAL 1 HOUR),'%Y-%m-%d %H:01:00') as crtdx from tm_dpo";
+	$logs=fetch_alla(exec_qry($conn,$sql));
+
+	for($cntr=0;$cntr<count($logs);$cntr++){
+		$data=$logs[$cntr];
+		$i=$data['oid'];
+		$h=$data['oname'];
+		$typ='offline';
+		$usr='dvo';
+		$rid=date("YmdHis").$cntr;
+		$d=$data['stts'];
+		$dt=$data['crtd'];
+		$st='wifi station';
+		$tm=$data['tm'];
+		$grp='link';
+		$ctr=$data['ctr'];
+		$stts='';
+		$p='';
+		if($ctr==1){
+			if($d=='Up'){
+				$closed='NOW()';
+				$stts='closed';
+				$p='Wifi Normal';
+				if(intval(date("H"))>9) $dt=$data['crtdx'];
+			}
+		}else{
+			$stts='new';
+			$closed='NULL';
+		}
+		
+		if($ctr==2||$stts=='closed'){
+			$sql="select oname,kanwil,lnk,wibstart,wibend,subtime(wibend,'2:30:00') as wibendsat from tm_outlets where oid='$i' and kanwil not in (select kanwil from tm_holidays where dt=DATE(NOW()))";
+			$ar=fetch_all(exec_qry($conn,$sql));
+			if(count($ar)>0){//not holiday
+				$stime="";
+				$etime="";
+				$found=false;
+				if(count($ar)>0){
+					$h=$ar[0][0];
+					$k=$ar[0][1];
+					//$st=$ar[0][2];
+					$stime=$ar[0][3];
+					$etime=strtolower(date('D'))=='sat'?$ar[0][5]:$ar[0][4];
+					$found=true;
+				}
+				if($found && $tm>=$stime && $tm<=$etime){ //working hour
+					$nows="DATE_FORMAT(NOW(),'%Y-%m-%d %H:%i:00')";
+					$nows="NOW()";
+					$sql="insert into tm_tickets (rowid,ticketno,dtm,createdby,lastupd,updby,dt,i,h,d,k,st,typ,grp,s,closed,p) values ('$rid','$rid',$nows,'$usr',$nows,'$usr','$dt','$i','$h','$d','$k','$st','$typ','$grp','$stts',$closed,'$p')";
+					$rs=exec_qry($conn,$sql);
+					$sql="delete from tm_dpo where oid='$i'";
+					$rs=exec_qry($conn,$sql);
+					if($stts=='closed') $rs=exec_qry($conn,"insert into tm_notes (ticketid,notes,s,lastupd,updby) values ('$rid','Up','$stts',now(),'$usr')");
+				}
+			}
+		}
+	}
 	
 	///solve ticket
-	$logs=array();
+	/*$logs=array();
 	$sql="SELECT ticketno FROM tm_tickets WHERE s<>'closed' AND s<>'solved' AND st='wifi station' AND i IN (SELECT oid from tm_dpo WHERE stts='Up')";
 	$logs=fetch_alla(exec_qry($conn,$sql));
 	for($cntr=0;$cntr<count($logs);$cntr++){
@@ -156,7 +230,7 @@ for($i=0;$i<count($data);$i++){
 		$rs=exec_qry($conn,$sql);
 		$sql="update tm_tickets set s='solved', lastupd=now(), updby='$usr' where ticketno='$ticketno'";
 		$rs=exec_qry($conn,$sql);
-	}
+	}*/
 
 
 }//end if holiday
